+++ checkpoint/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - `ocdbt_merge` option and unused `restore_with_serialized_types` option from
 `PyTreeCheckpointHandler`.
+- OCDBT coordinator code. These functions are no longer needed.
 
 ## [0.5.10] - 2024-04-22